requirements.txt
smartapi-python
pandas
requests
python-telegram-bot==13.7
pyotp
